requirements.txt
pytest
selenium
allure-pytest
Faker
assertpy